backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
matchering
numpy
scipy
soundfile
resampy
statsmodels
Flask
flask-cors
PyJWT
cryptography
python-dotenv
gunicorn
google-cloud-storage
google-cloud-firestore
google-cloud-pubsub
google-auth
requests